Moved applytuple and tohex to their own header file, also moved some public to private in patterntableview
