+++ src/Cpu6502.cpp
@@ -2,6 +2,7 @@
 #include <algorithm>
 #include <sstream>
 #include "Cpu6502.hpp"
+#include "functions.hpp" // toHex()
 
 #define EXECOPCODE(instrPtr, adringPtr) (this->*(instrPtr))((adringPtr))
 #define EXECADDRESSING(adringPtr) (this->*(adringPtr))()
@@ -511,14 +512,6 @@ inline uint8_t Cpu6502::POP() {
 /// ---------------- Opcode Functions ----------------
 ///
 ///
-
-template<typename T>
-std::string toHex(T&& num) {
-    static_assert (std::is_integral<typename std::decay<T>::type>::value, "must be an integral type");
-    std::stringstream stream;
-    stream << "0x" << std::hex << static_cast<int>(std::forward<T>(num));
-    return stream.str();
-}
 
 [[ noreturn ]]
 void Cpu6502::OP_ILLEGAL(AddressingPtr&) {
